Changed show class to default episode number to GUID
In cases where an episode number wasnt provided, the add-on would assign an incremental one. This
makes it difficult to match My TV sporting events that do not provide an episode number. Making
the GUID a base case makes it unique so it can be used as an alternate key.

Changed the Episode number field in the database from numeric to text.

Added necessary changes to infoLabels to reflect this change.

Added necessary changes to MyTV class to utilize this new base case.

Turned on link verification.

Version bump