PVR/EPG Tweaks.
As of this update, IPTV Simple Starts of disabled until the Slinger service is fully started. Then it is enabled.
On Kodi shutdown, IPTV Simple is once again disabled. This is to give Slinger enough time to prepare the guide files
for IPTV Simple. As a side-effect navigation to the Guide must happen otherwise Kodi thinks no PVR is installed if
there is only one. A user will see a screen flash otherwise it's seamless. The alternative is to bring up the guide
by default.

Also, Slinger no longer toggles IPTV enable state unless a forced update is requested. In it's default operation it
will simply update the guide and rely on IPTV Simple to pull in the changes.

Interval tweaks.

Version bump.