Reintroduced guide speed enhancement. There was a bug where the database connection was closed.
As a result the database would remain locked until the guide update was finished. Navigation would not be possible until the update finished (Android Phone had a hard crash for database lock.) This speed enhancement shows about a 36% improvement on average.
Same enhancement was introduced into the channel class for on-now.
The on-demand enhancement was removed, unable to verify source of 'ribbon_order' in channel 'metadata'. Result was breaking on-demand.
Version bump